@@ -2,10 +2,9 @@
   "name": "@11ty/eleventy-plugin-vite",
   "version": "4.0.0",
   "description": "A plugin to use Vite as a development server and run Vite to postprocess your Eleventy build.",
-  "main": ".eleventy.js",
   "license": "MIT",
   "engines": {
-    "node": ">=14.18.0"
+    "node": ">=18"
   },
   "funding": {
     "type": "opencollective",
@@ -16,7 +15,7 @@
     "server"
   ],
   "11ty": {
-    "compatibility": ">=2.0.0-canary.32"
+    "compatibility": ">=3.0.0-alpha.6"
   },
   "publishConfig": {
     "access": "public"
@@ -32,8 +31,28 @@
   },
   "bugs": "https://github.com/11ty/eleventy-plugin-vite/issues",
   "homepage": "https://github.com/11ty/eleventy-plugin-vite/",
+  "main": "./.eleventy.js",
+  "type": "module",
+  "exports": {
+    ".": "./.eleventy.js",
+    "./EleventyVite": "./EleventyVite.js",
+    "./package.json": "./package.json"
+  },
+  "files": [
+    ".eleventy.js",
+    "EleventyVite.js"
+  ],
+  "scripts": {
+    "example": "npx @11ty/eleventy --config=example/.eleventy.js --input=example --output=example/_site",
+    "example:start": "npm run example -- --serve",
+    "example:build": "npm run example",
+    "example:clean": "rimraf ./example/_site"
+  },
   "dependencies": {
-    "lodash.merge": "^4.6.2",
-    "vite": "^4.0.4"
+    "@11ty/eleventy-utils": "^1.0.3",
+    "vite": "^5.3.0"
+  },
+  "devDependencies": {
+    "@11ty/eleventy": "3.0.0-alpha.6"
   }
 }
