@@ -1,10 +1,21 @@
-const { promises: fsp } = require("fs");
-const path = require("path");
-const lodashMerge = require("lodash.merge");
+import { promises as fsp } from "node:fs";
+import path from "node:path";
+import { DeepCopy, Merge } from "@11ty/eleventy-utils";
+import { build, createServer } from "vite";
 
+/** @type {Required<import(".eleventy.js").EleventyViteOptions>} */
 const DEFAULT_OPTIONS = {
   tempFolderName: ".11ty-vite",
   viteOptions: {
+    clearScreen: false,
+    appType: "mpa",
+    server: {
+      middlewareMode: true,
+    },
+    build: {
+      emptyOutDir: true,
+      rollupOptions: {}, // we use this to inject input for MPA build below
+    },
     resolve:{
       alias:{
         // Allow references to `node_modules` directly for bundling.
@@ -12,30 +23,26 @@ const DEFAULT_OPTIONS = {
         // Note that bare module specifiers are also supported
       },
     },
-    clearScreen: false,
-    appType: "mpa",
-    server: {
-      mode: "development",
-      middlewareMode: true,
-    },
-    build: {
-      mode: "production",
-      rollupOptions: {}, // we use this to inject input for MPA build below
-    }
   }
 }
 
-class EleventyVite {
-  constructor(outputDir, pluginOptions = {}) {
-    this.outputDir = outputDir;
-    this.options = lodashMerge({}, DEFAULT_OPTIONS, pluginOptions);
+export default class EleventyVite {
+  /** @type {import("@11ty/eleventy/src/Util/ProjectDirectories.js").default} */
+  directories;
+
+  /** @type {Required<import(".eleventy.js").EleventyViteOptions>} */
+  options;
+
+  constructor(directories, pluginOptions = {}) {
+    this.directories = directories;
+    this.options = Merge({}, DEFAULT_OPTIONS, pluginOptions);
   }
 
   async getServerMiddleware() {
-    let viteOptions = lodashMerge({}, this.options.viteOptions);
-    viteOptions.root = this.outputDir;
+    /** @type {import("vite").InlineConfig} */
+    let viteOptions = DeepCopy({}, this.options.viteOptions);
+    viteOptions.root = this.directories.output;
 
-    const { createServer } = await import('vite');
     let vite = await createServer(viteOptions);
 
     return vite.middlewares;
@@ -46,33 +53,32 @@ class EleventyVite {
   }
 
   async runBuild(input) {
-    let tmp = path.resolve(".", this.options.tempFolderName);
+    let tmp = path.resolve(this.directories.input, this.options.tempFolderName);
 
-    await fsp.mkdir(tmp, { recursive: true });
-    await fsp.rename(this.outputDir, tmp);
+    await fsp.rename(this.directories.output, tmp);
 
     try {
-      let viteOptions = lodashMerge({}, this.options.viteOptions);
+      /** @type {import("vite").InlineConfig} */
+      let viteOptions = DeepCopy({}, this.options.viteOptions);
       viteOptions.root = tmp;
 
       viteOptions.build.rollupOptions.input = input
         .filter(entry => !!entry.outputPath) // filter out `false` serverless routes
         .filter(entry => (entry.outputPath || "").endsWith(".html")) // only html output
         .map(entry => {
-          if(!entry.outputPath.startsWith(this.outputDir + path.sep)) {
-            throw new Error(`Unexpected output path (was not in output directory ${this.outputDir}): ${entry.outputPath}`);
+          if(!entry.outputPath.startsWith(this.directories.output)) {
+            throw new Error(`Unexpected output path (was not in output directory ${this.directories.output}): ${entry.outputPath}`);
           }
 
-          return path.resolve(tmp, entry.outputPath.substr(this.outputDir.length + path.sep.length));
+          return path.resolve(tmp, entry.outputPath.substring(this.directories.output.length));
         });
 
-      viteOptions.build.outDir = path.resolve(".", this.outputDir);
+      viteOptions.build.outDir = path.resolve(".", this.directories.output);
 
-      const { build } = await import('vite');
       await build(viteOptions);
     } catch(e) {
-      console.warn( `[11ty] Encountered a Vite build error, restoring original Eleventy output to ${this.outputDir}`, e );
-      await fsp.rename(tmp, this.outputDir);
+      console.warn( `[11ty] Encountered a Vite build error, restoring original Eleventy output to ${this.directories.output}`, e );
+      await fsp.rename(tmp, this.directories.output);
       throw e;
     } finally {
       // remove the tmp dir
@@ -81,4 +87,3 @@ class EleventyVite {
   }
 }
 
-module.exports = EleventyVite;
